+++ frameworks/Rust/actix/benchmark_config.json
@@ -1,7 +1,7 @@
 {
-  "framework": "actix",
+  "framework": "Actix Web",
   "tests": [{
-    "default": {
+    "web": {
       "json_url": "/json",
       "plaintext_url": "/plaintext",
       "port": 8080,
@@ -11,53 +11,15 @@
       "framework": "actix",
       "language": "Rust",
       "orm": "Raw",
-      "platform": "None",
+      "platform": "actix-http",
       "webserver": "actix-web",
       "os": "Linux",
       "database_os": "Linux",
-      "display_name": "Actix",
+      "display_name": "Actix Web [Postgres]",
       "notes": "",
-      "versus": ""
+      "versus": "actix-http"
     },
-    "core": {
-      "fortune_url": "/fortunes",
-      "db_url": "/db",
-      "query_url": "/queries?q=",
-      "update_url": "/updates?q=",
-      "port": 8080,
-      "approach": "Realistic",
-      "classification": "Platform",
-      "database": "Postgres",
-      "framework": "actix",
-      "language": "Rust",
-      "orm": "Raw",
-      "platform": "None",
-      "webserver": "actix-web",
-      "os": "Linux",
-      "database_os": "Linux",
-      "display_name": "Actix [Platform]",
-      "notes": "",
-      "versus": ""
-    },
-    "raw": {
-      "json_url": "/json",
-      "plaintext_url": "/plaintext",
-      "port": 8080,
-      "approach": "Realistic",
-      "classification": "Platform",
-      "database": "Postgres",
-      "framework": "actix",
-      "language": "Rust",
-      "orm": "Raw",
-      "platform": "None",
-      "webserver": "actix-web",
-      "os": "Linux",
-      "database_os": "Linux",
-      "display_name": "Actix [Raw]",
-      "notes": "",
-      "versus": ""
-    },
-    "diesel": {
+    "web-diesel": {
       "db_url": "/db",
       "fortune_url": "/fortunes",
       "query_url": "/queries?q=",
@@ -69,31 +31,49 @@
       "framework": "actix",
       "language": "Rust",
       "orm": "Full",
-      "platform": "None",
+      "platform": "actix-http",
       "webserver": "actix-web",
       "os": "Linux",
       "database_os": "Linux",
-      "display_name": "Actix [Diesel]",
+      "display_name": "Actix Web [Diesel]",
       "notes": "",
-      "versus": ""
+      "versus": "actix-http"
     },
-    "pg": {
+    "http": {
       "db_url": "/db",
       "fortune_url": "/fortunes",
       "query_url": "/queries?q=",
       "update_url": "/updates?q=",
       "port": 8080,
       "approach": "Realistic",
-      "classification": "Micro",
+      "classification": "Platform",
       "database": "Postgres",
       "framework": "actix",
       "language": "Rust",
       "orm": "Raw",
       "platform": "None",
-      "webserver": "actix-web",
+      "webserver": "actix-http",
       "os": "Linux",
       "database_os": "Linux",
-      "display_name": "Actix [Postgres]",
+      "display_name": "actix-http",
+      "notes": "",
+      "versus": "actix-server"
+    },
+    "server": {
+      "json_url": "/json",
+      "plaintext_url": "/plaintext",
+      "port": 8080,
+      "approach": "Stripped",
+      "classification": "Platform",
+      "database": "Postgres",
+      "framework": "actix",
+      "language": "Rust",
+      "orm": "Raw",
+      "platform": "None",
+      "webserver": "actix-server",
+      "os": "Linux",
+      "database_os": "Linux",
+      "display_name": "actix-server",
       "notes": "",
       "versus": ""
     }
