+++ frameworks/Rust/actix/rustfmt.toml
@@ -1,2 +1,2 @@
-max_width = 89
+max_width = 96
 reorder_imports = true
