+++ frameworks/Rust/actix/src/main_server.rs
@@ -8,24 +8,26 @@ use std::{
     task::{Context, Poll},
 };
 
-use actix_codec::{AsyncRead, AsyncWrite, Decoder};
+use actix_codec::{AsyncWrite, Decoder};
 use actix_http::{h1, Request};
 use actix_server::Server;
 use actix_service::fn_service;
 use bytes::{Buf, BufMut, BytesMut};
 use simd_json_derive::Serialize;
-use tokio::{io::AsyncBufRead, net::TcpStream};
+use tokio::net::TcpStream;
 
 mod models;
 mod utils;
 
 use crate::utils::Writer;
 
-const JSON: &[u8] = b"HTTP/1.1 200 OK\r\nServer: A\r\nContent-Type: application/json\r\nContent-Length: 27\r\n";
-const PLAIN: &[u8] = b"HTTP/1.1 200 OK\r\nServer: A\r\nContent-Type: text/plain\r\nContent-Length: 13\r\n";
-const HTTPNFOUND: &[u8] = b"HTTP/1.1 400 OK\r\n";
-const HDR_SERVER: &[u8] = b"Server: A\r\n";
-const BODY: &[u8] = b"Hello, World!";
+const JSON_HEAD: &[u8] =
+    b"HTTP/1.1 200 OK\r\nServer: A\r\nContent-Type: application/json\r\nContent-Length: 27\r\n";
+const PLAIN_HEAD: &[u8] =
+    b"HTTP/1.1 200 OK\r\nServer: A\r\nContent-Type: text/plain\r\nContent-Length: 13\r\n";
+const NOT_FOUND_HEAD: &[u8] = b"HTTP/1.1 204 OK\r\nServer: A\r\n";
+const END_HEADERS: &[u8] = b"\r\n";
+const PLAIN_BODY: &[u8] = b"Hello, World!";
 
 #[derive(Serialize)]
 pub struct Message {
@@ -46,22 +48,26 @@ impl App {
                 let message = Message {
                     message: "Hello, World!",
                 };
-                self.write_buf.put_slice(JSON);
-                self.codec.config().set_date(&mut self.write_buf);
+                self.write_buf.put_slice(JSON_HEAD);
+                self.codec.config().write_date_header(&mut self.write_buf, false);
+                self.write_buf.put_slice(END_HEADERS);
+
                 message
                     .json_write(&mut Writer(&mut self.write_buf))
                     .unwrap();
             }
 
             "/plaintext" => {
-                self.write_buf.put_slice(PLAIN);
-                self.codec.config().set_date(&mut self.write_buf);
-                self.write_buf.put_slice(BODY);
-            }
+                self.write_buf.put_slice(PLAIN_HEAD);
+                self.codec.config().write_date_header(&mut self.write_buf, false);
+                self.write_buf.put_slice(END_HEADERS);
 
+                self.write_buf.put_slice(PLAIN_BODY);
+            }
+            
             _ => {
-                self.write_buf.put_slice(HTTPNFOUND);
-                self.write_buf.put_slice(HDR_SERVER);
+                self.write_buf.put_slice(NOT_FOUND_HEAD);
+                self.write_buf.put_slice(END_HEADERS);
             }
         }
     }
@@ -74,20 +80,23 @@ impl Future for App {
         let this = self.get_mut();
 
         loop {
+            if this.io.poll_read_ready(cx).is_pending() {
+                break;
+            }
+            
             if this.read_buf.capacity() - this.read_buf.len() < 512 {
                 this.read_buf.reserve(32_768);
             }
 
-            let read = Pin::new(&mut this.io).poll_read_buf(cx, &mut this.read_buf);
+            let io = Pin::new(&mut this.io);
 
-            match read {
-                Poll::Pending => break,
-                Poll::Ready(Ok(n)) => {
-                    if n == 0 {
-                        return Poll::Ready(Ok(()));
-                    }
-                }
-                Poll::Ready(Err(_)) => return Poll::Ready(Err(())),
+            match io.try_read_buf(&mut this.read_buf) {
+                Ok(0) => return Poll::Ready(Ok(())),
+
+                Ok(_) => {}
+                Err(ref err) if err.kind() == io::ErrorKind::WouldBlock => break,
+                
+                Err(_) => return Poll::Ready(Err(())),
             }
         }
 
@@ -109,15 +118,10 @@ impl Future for App {
 
             while written < len {
                 match Pin::new(&mut this.io).poll_write(cx, &this.write_buf[written..]) {
-                    Poll::Pending => {
-                        break;
-                    }
+                    Poll::Pending => break,
+                    Poll::Ready(Ok(0)) => return Poll::Ready(Ok(())),
                     Poll::Ready(Ok(n)) => {
-                        if n == 0 {
-                            return Poll::Ready(Ok(()));
-                        } else {
-                            written += n;
-                        }
+                        written += n;
                     }
                     Poll::Ready(Err(_)) => return Poll::Ready(Err(())),
                 }
@@ -130,18 +134,24 @@ impl Future for App {
             }
         }
 
+        match Pin::new(&mut this.io).poll_flush(cx) {
+            Poll::Pending => {}
+            Poll::Ready(Ok(())) => {}
+            Poll::Ready(Err(_)) => return Poll::Ready(Err(())),
+        }
+
         Poll::Pending
     }
 }
 
 #[actix_web::main]
 async fn main() -> io::Result<()> {
-    println!("Started http server: 127.0.0.1:8080");
+    eprintln!("Started http server: 127.0.0.1:8080");
 
     // start http server
     Server::build()
         .backlog(1024)
-        .bind("techempower", "0.0.0.0:8080", || {
+        .bind("tfb-actix-server", "0.0.0.0:8080", || {
             fn_service(|io: TcpStream| App {
                 io,
                 read_buf: BytesMut::with_capacity(32_768),
