+++ frameworks/Rust/actix/src/main_web.rs
@@ -6,19 +6,18 @@ use std::time::Duration;
 use actix_http::{HttpService, KeepAlive};
 use actix_service::map_config;
 use actix_web::{
-    body,
     dev::{AppConfig, Server},
     http::{
         header::{HeaderValue, CONTENT_TYPE, SERVER},
         StatusCode,
     },
-    web, App, HttpResponse,
+    web::{self, Bytes, BytesMut},
+    App, HttpResponse,
 };
-use bytes::{Bytes, BytesMut};
 use simd_json_derive::Serialize;
 
 mod utils;
-use utils::{Writer, SIZE};
+use utils::{Writer, JSON_MSG_SIZE};
 
 #[derive(Serialize)]
 pub struct Message {
@@ -29,7 +28,8 @@ async fn json() -> HttpResponse<Bytes> {
     let message = Message {
         message: "Hello, World!",
     };
-    let mut body = BytesMut::with_capacity(SIZE);
+
+    let mut body = BytesMut::with_capacity(JSON_MSG_SIZE);
     message.json_write(&mut Writer(&mut body)).unwrap();
 
     let mut res = HttpResponse::with_body(StatusCode::OK, body.freeze());
@@ -40,11 +40,8 @@ async fn json() -> HttpResponse<Bytes> {
     res
 }
 
-async fn plaintext() -> HttpResponse {
-    let mut res = HttpResponse::with_body(
-        StatusCode::OK,
-        Body::Bytes(Bytes::from_static(b"Hello, World!")),
-    );
+async fn plaintext() -> HttpResponse<Bytes> {
+    let mut res = HttpResponse::with_body(StatusCode::OK, Bytes::from_static(b"Hello, World!"));
     res.headers_mut()
         .insert(SERVER, HeaderValue::from_static("A"));
     res.headers_mut()
@@ -54,15 +51,15 @@ async fn plaintext() -> HttpResponse {
 
 #[actix_web::main]
 async fn main() -> std::io::Result<()> {
-    println!("Started http server: 127.0.0.1:8080");
+    println!("Started HTTP server: 127.0.0.1:8080");
 
     // start http server
     Server::build()
         .backlog(1024)
-        .bind("techempower", "0.0.0.0:8080", || {
+        .bind("tfb-actix-web", "0.0.0.0:8080", || {
             HttpService::build()
                 .keep_alive(KeepAlive::Os)
-                .client_timeout(Duration::ZERO)
+                .client_request_timeout(Duration::ZERO)
                 .h1(map_config(
                     App::new()
                         .service(web::resource("/json").to(json))
